# docs/swagger.json
{
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API Server for films and actors",
        "title": "Filmlib API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8091",
    "basePath": "/",
    "paths": {
        "/candidate/auth": {
            "post": {
                "description": "login",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "SignIn",
                "operationId": "login",
                "parameters": [
                    {
                        "description": "credentials",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.userAuthDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "jwt",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Failed to decode request body. Invalid JSON",
                        "schema": {}
                    },
                    "500": {
                        "description": "Failed to generate JWT",
                        "schema": {}
                    }
                }
            }
        },
        "/candidate/registration": {
            "post": {
                "description": "create account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Registration",
                "operationId": "create-account",
                "parameters": [
                    {
                        "description": "account info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.userRegDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    },
                    "default": {
                        "description": "",
                        "schema": {}
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.userAuthDTO": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "phonenumber": {
                    "type": "string"
                }
            }
        },
        "handler.userRegDTO": {
            "type": "object",
            "properties": {
                "phonenumber": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}